+++ examples/java/DonutIO/WinStage.java
@@ -7,7 +7,9 @@ import org.openpatch.scratch.extensions.text.Text;
 
 public class WinStage extends Stage {
   public WinStage() {
-    this.add(new Background());
+    var bg = new Background();
+    bg.setTransparency(50);
+    this.add(bg);
 
     var text = new Text();
     text.setTextSize(48);