+++ examples/java/Sensing/Sensing.java
@@ -17,10 +17,22 @@ public class Sensing extends Stage {
     m = new MovableHero();
     this.add(h);
     this.add(m);
+
+    var uiH = new Hero();
+    uiH.isUI(true);
+    uiH.setPosition(300, 300);
+    this.add(uiH);
   }
 
   public void run() {
     this.display("Move the hero with WASD and rotate him with R");
+
+    if(isKeyPressed(KeyCode.VK_0)) {
+      this.getCamera().changeZoom(-1);
+    }
+    if (isKeyPressed(KeyCode.VK_1)) {
+      this.getCamera().changeZoom(1);
+    }
   }
 
   public static void main(String[] args) {
@@ -77,7 +89,7 @@ class MovableHero extends Hero {
     if (this.isKeyPressed(82)) {
       this.turnRight(1);
     }
-    if (this.isTouchingSprite(Sensing.h)) {
+    if (this.isTouchingSprite(Hero.class)) {
       this.say("Hit");
     } else {
       this.say(null);