+++ CHANGELOG.md
@@ -3,6 +3,11 @@ name: Changelog
 index: 4
 ---
 
+## 4.6.0
+
+- 🚀 Feat: Add setCursor method to the stage class.
+- 🚀 Feat (EXPERIMENTAL): Add extension camera. The camera object allows you to move the view of the stage without manipulating the coordinates of e.g. sprites.
+
 ## 4.5.0
 
 - 💥 BREAKING CHANGE: Only use double for methods instead of supporting both float and doubles. This was unnecessary and cluttered the BlueJ interface with "duplicated" methods. This could potentially break your project if you were using floats, just replace them with doubles, and you are good to go.